requirements.txt
python-telegram-bot>=20.0
Flask==2.3.2
APScheduler==3.6.3